theora filter メモリリーク修正、フォーマット誤認修正、警告修正、不要ファイル削除。

+++ visual/win32/krmovie/ogg/AutoOggChainGranuleSeekTable.cpp
@@ -43,7 +43,7 @@ bool AutoOggChainGranuleSeekTable::buildTable()
 		unsigned char* locBuff = new unsigned char[BUFF_SIZE];		//Deleted this function.
 		while (!mFile.eof()) {
 			mFile.read((char*)locBuff, BUFF_SIZE);
-			mOggDemux->feed((const unsigned char*)locBuff, mFile.gcount());
+			mOggDemux->feed((const unsigned char*)locBuff, (unsigned long)mFile.gcount());
 		}
 		delete[] locBuff;
